third_party/WebKit/LayoutTests/http/tests/security/preload-script-crossorigin-fails-cross-origin-expected.txt

CONSOLE ERROR: Script from origin 'http://localhost:8000' has been blocked from loading by Cross-Origin Resource Sharing policy: No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
Preload of script with CORS failure only shows one error

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Did preload.
PASS Script error.
PASS successfullyParsed is true

TEST COMPLETE
